refactor: remove circular dependencies in rest of packages
