+++ packages/picasso-charts/src/LineChart/index.ts
@@ -3,6 +3,6 @@ import { OmitInternalProps } from '@toptal/picasso-shared'
 import { Props } from './LineChart'
 
 export { default } from './LineChart'
-export type { HighlightConfig, ReferenceLineType } from './LineChart'
+export type { ReferenceLineType } from './LineChart'
 
 export type LineChartProps = OmitInternalProps<Props>
