+++ packages/picasso-forms/src/types.ts
@@ -1,0 +1,15 @@
+import { DateOrDateRangeType } from '@toptal/picasso'
+import { Item } from '@toptal/picasso/Autocomplete'
+import { FileUpload } from '@toptal/picasso/FileInput'
+
+export type ValueType =
+  | string
+  | string[]
+  | number
+  | boolean
+  | null
+  | undefined
+  | FileUpload[]
+  | DateOrDateRangeType
+  | Item
+  | Item[]
